pinn: move files
